some changes

+++ include/task2.h
@@ -2,11 +2,11 @@
 #include <iostream>
 using namespace std;
 
-template < typename T, size_t N >
+template < typename T, int N >
 
 T* createArr(T(*gen)()) {
 	T* temp = new T[N];
-	for (size_t i = 0; i < N; i++)
+	for (int i = 0; i < N; i++)
 		temp[i] = gen();
 	return temp;
 }
