+++ include/task3.h
@@ -1,6 +1,6 @@
 #pragma once
-template <typename T, size_t N>
+template <typename T, int N>
 void map(T* arr, T(*change)()) {
-	for (size_t i = 0; i < N; i++)
+	for (int i = 0; i < N; i++)
 		arr[i] = change();
 }
