+++ src/main2.cpp
@@ -8,9 +8,9 @@ T gen() {
 
 
 int main() {
-	const size_t N = 10;
-	size_t* mas;
-	mas = createArr<size_t, N>(gen);
-	for (size_t i = 0; i < N; i++)
+	const int N = 10;
+	int* mas;
+	mas = createArr<int, N>(gen);
+	for (int i = 0; i < N; i++)
 		cout << mas[i] << " ";
 }
